herbarium_project/requirements.txt
pillow
pandas
psycopg2-binary
